+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Models\Category;
 use App\Models\SubCategory;
 use Str;
+use App\Models\ChildCategory;
 
 class SubCategoryController extends Controller
 {
@@ -99,6 +100,12 @@ class SubCategoryController extends Controller
     public function destroy(string $id)
     {
         $subCategory = SubCategory::findOrFail($id);
+        $childCategory = ChildCategory::where('sub_category_id', $subCategory->id)->count();
+
+        if($childCategory > 0){
+            return response(['status' => 'error', 'message' => 'This items contain, sub items for delete this you have to delete the sub items first!']);
+        }
+
         $subCategory->delete();
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
     }
